Add JSON schema to `Pydantic.ByteSize` through annotation
Adding JSON schema to `Pydantic.ByteSize`
through annotation instead of subclassing
allows objects of type `Pydantic.ByteSize`
to assign to a `ByteSizeJsonSchema` attribute
without static type checker complain
