+++ dandischema/models.py
@@ -9,10 +9,8 @@ from pydantic import (
     UUID4,
     AnyHttpUrl,
     BaseModel,
-    ByteSize,
     EmailStr,
     Field,
-    GetCoreSchemaHandler,
     GetJsonSchemaHandler,
     TypeAdapter,
     ValidationInfo,
@@ -20,41 +18,13 @@ from pydantic import (
     model_validator,
 )
 from pydantic.json_schema import JsonSchemaValue
-from pydantic_core import CoreSchema, core_schema
-from typing_extensions import Annotated
+from pydantic_core import CoreSchema
 
 from .consts import DANDI_SCHEMA_VERSION
 from .digests.dandietag import DandiETag
 from .digests.zarr import ZARR_CHECKSUM_PATTERN, parse_directory_digest
+from .types import ByteSizeJsonSchema
 from .utils import name2title
-
-
-class _ByteSizeJsonSchemaAnnotation:
-    """
-    An annotation for `pydantic.ByteSize` that provides a JSON schema
-
-    Note: Pydantic V2 doesn't provide a JSON schema for `pydantic.ByteSize`.
-          This subclass provides a JSON schema that is the same JSON schema
-          used for `pydantic.ByteSize` in Pydantic V1.
-    """
-
-    @classmethod
-    def __get_pydantic_core_schema__(
-        cls, source: Type[ByteSize], handler: GetCoreSchemaHandler
-    ) -> CoreSchema:
-        assert source is ByteSize
-        return handler(source)
-
-    @classmethod
-    def __get_pydantic_json_schema__(
-        cls,
-        _core_schema: CoreSchema,
-        handler: GetJsonSchemaHandler,
-    ) -> JsonSchemaValue:
-        return handler(core_schema.int_schema())
-
-
-ByteSizeJsonSchema = Annotated[ByteSize, _ByteSizeJsonSchemaAnnotation()]
 
 # Use DJANGO_DANDI_WEB_APP_URL to point to a specific deployment.
 DANDI_INSTANCE_URL: Optional[str]
